test(architecture): split semantic value dependency check
