+++ internal/platform/architecture/architecture_test.go
@@ -155,17 +155,6 @@ func TestSemanticValueIsAPublishedAnalyticsDependency(t *testing.T) {
 	}
 }
 
-func TestSemanticValueIsAPublishedAccessDependency(t *testing.T) {
-	source, sourceOK := ClassifyPackage("internal/access")
-	target, targetOK := ClassifyPackage("internal/semanticvalue")
-	if !sourceOK || !targetOK {
-		t.Fatalf("classify access=%v semanticvalue=%v", sourceOK, targetOK)
-	}
-	if violation := CapabilityImportViolation("internal/access", source, "internal/semanticvalue", target); violation != "" {
-		t.Fatalf("access -> semanticvalue violation = %q, want published contract", violation)
-	}
-}
-
 func TestAgentGeneratedAPIIsCapabilityOwned(t *testing.T) {
 	rule, ok := ClassifyPackage("internal/agent/api/gen")
 	if !ok {
